+++ e2e/azurerm_estimation_test.go
@@ -11,6 +11,7 @@ import (
 	"github.com/cycloidio/terracost/cost"
 	"github.com/cycloidio/terracost/mysql"
 	"github.com/cycloidio/terracost/testutil"
+	"github.com/cycloidio/terracost/usage"
 	"github.com/shopspring/decimal"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -44,7 +45,7 @@ func TestAzureRMEstimation(t *testing.T) {
 		require.NoError(t, err)
 		defer f.Close()
 
-		plan, err := costestimation.EstimateTerraformPlan(ctx, backend, f)
+		plan, err := costestimation.EstimateTerraformPlan(ctx, backend, f, usage.Default)
 		require.NoError(t, err)
 
 		pcost, err := plan.PriorCost()
@@ -56,7 +57,7 @@ func TestAzureRMEstimation(t *testing.T) {
 		assertCostEqual(t, cost.NewMonthly(decimal.NewFromFloat(64.021), "USD"), pcost)
 	})
 	t.Run("FromHCL", func(t *testing.T) {
-		plan, err := costestimation.EstimateHCL(ctx, backend, nil, "../testdata/azurerm/stack-compute")
+		plan, err := costestimation.EstimateHCL(ctx, backend, nil, "../testdata/azurerm/stack-compute", usage.Default)
 		require.NoError(t, err)
 
 		assert.Nil(t, plan.Prior)